+++ drivers/lora/CMakeLists.txt
@@ -3,4 +3,5 @@
 zephyr_library_named(loramac-node)
 
 zephyr_library_sources_ifdef(CONFIG_LORA_SHELL shell.c)
+zephyr_library_sources_ifdef(CONFIG_HAS_SEMTECH_RADIO_DRIVERS hal_common.c)
 zephyr_library_sources_ifdef(CONFIG_LORA_SX1276 sx1276.c)
